chore: tidy up comments, tests and specs
