+++ svc/api/openapi/spec/paths/v2/analytics/getGatewayRequests/index.yaml
@@ -10,7 +10,7 @@ post:
     A query can use only these four public aliases: `gateway_requests_v1`, `gateway_requests_per_minute_v1`, `gateway_requests_per_hour_v1`, or `gateway_requests_per_day_v1`. CTEs, subqueries, UNION, and EXCEPT are permitted.
     The root key must have the `project.*.read_analytics` permission.
     Unkey limits each query to the workspace of the root key. To get the data for one project, app, or environment, add a filter on `project_id`, `app_id`, or `environment_id`.
-    The raw table does not give access to the internal Unkey columns. The workspace retention period and the workspace query limits apply.
+    The workspace retention period and the workspace query limits apply.
     For the tables, the columns, and more query examples, see [Query gateway request analytics](/platform/analytics/get-gateway-requests).
   requestBody:
     required: true
